emacs: Remove remote-host notify from ERC
This hack was in place to use ssh to run notify-send on a remote host in
case Emacs/ERC was running somewhere other than the client (so that
running notify-send on the Emacs host was useless). This was never very
robust and basically never worked for work where I need it most, because
security. I'll have to figure something else out.